v 0.4.0 implemented (#5)
